refactor(stagewise): reorganize hotkey bindings, fix drag overlay, rewire shortcuts
- Split hotkeys into GlobalHotkeyBindings (always mounted), AgentHotkeyBindings
  (always mounted), and BrowserTabHotkeys (per-content-panel)
- Cmd+B now always works regardless of content panel state
- Cmd+T creates tabs attached to current agent and opens content panel
- Mod+PageDown/Up now immediately switches agents (same as Ctrl+Tab logic)
- Cmd+I toggles context selection, always focuses chat, never unfocuses
- Cmd+L toggles chat input focus, Cmd+Opt+L toggles omnibox focus
- Removed HOME_PAGE hotkey, RESTORE_TAB is stubbed (no backend yet)
- Drag overlay tabs now match regular tab styling (border radius, icon size,
  width, padding, no pin/close icons)
- Deleted CoreHotkeyBindings, renamed to BrowserTabHotkeys
